requirements.txt
python-dotenv
pydantic
fastapi
uvicorn
pymongo[srv]>=3.11
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
langchain
langchain-community
langchain-text-splitters
pymupdf
pinecone
ipykernel
python-multipart
certifi
langchain-groq